add a generic dimension_ordering and use it for tile_compare_xyz
After all, there are 48 possibilities
(3! = 6 orderings of x, y, z, times 2^3 = 8 sign combinations).
We wouldn't want to write them by hand.

But is it really worth the template madness?

I tried to use arrays (static const int dimensions[3] = { ... })
but C++ declined to let me do that, and it was not a big loss
for this particular code arrangement.