Calculate the manhattan distance for each polygon displayed.
I did it a bit wrong, and it displayed even more wrong than it should
for the wrongness I found.  But on the plus side, it didn't make
the simulation go slower (after all I was ripping out the other
manhattan distance calls at the same time), and I got to revert
the ugly tile-sides-ordering code that didn't quite work either.