fix bug introduced by 31008a40993
The other part of 31008a40993 was a correct bugfix,
but this part was solely a bug-introduction in some cases
of sqrt of max uint64.  The uint32+1 overflows there,
while upper_bound-1 never overflows because it is never
zero given correct type parameters to isqrt_impl.
(Given wrong type parameters, upper_bound could have overflowed
to zero earlier in the function.)