fix ostream_bundle
In Emscripten, the destructor of the locale object inside
the ios_base of the ostream the ostream_bundle was derived
from, called abort() for no obvious reason.  Luckily,
there was no reason for ostream_bundle to be deriving from
ostream in the first place, aside from a mistake in a non-member
operator's signature.