upgrade OpenGL to the modern shader era
The display still does the same thing.  It is just more
efficient with / compatible with modern hardware and software
(such as OpenGL ES 2.0).

Putting triangles instead of quads in the buffer takes
up somewhat more space, which could be mitigated by using
an OpenGL index buffer if it seemed like a priority.

There may still be some inefficiencies.